refactor(go): G7 — move embedded under typescript/ and retarget staging
