+++ packages/go/driver/internal/typescript/embedded/doc.go
