+++ packages/go/driver/internal/typescript/embedded/sidecar_darwin_amd64.go
