+++ packages/go/driver/internal/typescript/embedded/sidecar_darwin_arm64.go
