+++ packages/go/driver/internal/typescript/embedded/sidecar_linux_amd64.go
