+++ packages/go/driver/internal/typescript/embedded/sidecar_linux_arm64.go
